IDM-Gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
grpcio==1.68.0
grpcio-tools==1.68.0
protobuf==5.28.3
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
